tests/wpt/tests/css/css-fonts/parsing/font-stretch-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 4: getComputedStyle().fontStretch</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-4/#font-stretch-prop">
<meta name="assert" content="font-stretch computed value is a percentage.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value('font-stretch', 'ultra-condensed', '50%');
test_computed_value('font-stretch', 'extra-condensed', '62.5%');
test_computed_value('font-stretch', 'condensed', '75%');
test_computed_value('font-stretch', 'semi-condensed', '87.5%');
test_computed_value('font-stretch', 'normal', '100%');
test_computed_value('font-stretch', 'semi-expanded', '112.5%');
test_computed_value('font-stretch', 'expanded', '125%');
test_computed_value('font-stretch', 'extra-expanded', '150%');
test_computed_value('font-stretch', 'ultra-expanded', '200%');

test_computed_value('font-stretch', '234.5%');
</script>
</body>
</html>
